Refactor(outfit/mount): Attribute System and AttackSpeed fix
- Fix: Corrected attackSpeed modification logic in mounts.cpp and outfit.cpp to ensure values are added/removed correctly during state changes.
- Refactor: Moved CONDITIONID_OUTFIT and CONDITIONID_MOUNT to enums.h and exposed them to Lua via luascript.cpp.
- Scaling: Changed mount ID parameters from uint8_t to uint16_t across player.cpp, mounts.cpp, and related headers.
- Data: Updated mounts.xml and outfits.xml with new attribute examples (healthGain, manaGain, attackSpeed).

+++ src/mounts.h
@@ -52,8 +52,8 @@ public:
 	Mount* getMountByName(std::string_view name);
 	Mount* getMountByClientID(uint16_t clientId);
 
-	bool addAttributes(uint32_t playerId, uint8_t mountId);
-	bool removeAttributes(uint32_t playerId, uint8_t mountId);
+	bool addAttributes(uint32_t playerId, uint16_t mountId);
+	bool removeAttributes(uint32_t playerId, uint16_t mountId);
 
 	const std::vector<Mount>& getMounts() const { return mounts; }
