+++ src/player.h
@@ -141,7 +141,7 @@ public:
 	void setCurrentMount(uint16_t mountId);
 	bool isMounted() const { return defaultOutfit.lookMount != 0; }
 	bool toggleMount(bool mount);
-	bool changeMount(uint8_t mountId, bool checkList = true);
+	bool changeMount(uint16_t mountId, bool checkList = true);
 	bool tameMount(uint16_t mountId);
 	bool untameMount(uint16_t mountId);
 	bool hasMount(const Mount* mount) const;
